requirements.txt
attrs==19.3.0
beautifulsoup4==4.9.0
boto3==1.12.43
botocore==1.15.43
certifi==2020.4.5.1
cffi==1.14.0
chalice==1.13.1
chardet==3.0.4
click==7.1.1
cryptography==2.9
decorator==4.4.2
docutils==0.15.2
enum-compat==0.0.3
future==0.18.2
idna==2.9
jmespath==0.9.5
lxml==4.5.0
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==12.6.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
tornado==6.0.4
typing==3.6.4
urllib3==1.25.9
